+++ common/umbrella/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [2.24.1](https://github.com/Vonage/vivid/compare/v2.24.0...v2.24.1) (2021-12-23)
+
+
+
+# [2.24.0](https://github.com/Vonage/vivid/compare/v2.23.0...v2.24.0) (2021-12-22)
+
+
+
+# [2.23.0](https://github.com/Vonage/vivid/compare/v2.22.1...v2.23.0) (2021-12-14)
+
+
+
 ## [2.22.1](https://github.com/Vonage/vivid/compare/v2.22.0...v2.22.1) (2021-11-05)